pyproject.toml
[project]
name = "ertsi-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "pyyaml>=6.0.2",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
]
